+++ docusaurus.config.js
@@ -40,6 +40,7 @@ const siteConfig = {
   ],
 
   themeConfig: {
+    /*
     announcementBar: {
       id: 'shop',
       content:
@@ -47,6 +48,7 @@ const siteConfig = {
       backgroundColor: '#0091db',
       textColor: 'white',
     },
+    */
     colorMode: {
       defaultMode: 'light',
       disableSwitch: false,
@@ -217,7 +219,9 @@ const siteConfig = {
             ],
         },
         { to: "blog", label: "Blog", position: "left" },
+        /*
         { href: "https://shop.vhdplus.com", label: "Shop", position: "left" },
+        */
         { to: "/docs/contact", label: "Contact", position: "right" },
         {
           href: 'https://www.youtube.com/channel/UC7qiOvlaBSiWyAb7R1xTaEw',